+++ extensions/brew/src/brew.ts
@@ -293,6 +293,10 @@ export async function brewUpgradeAll(greedy: boolean, cancel?: AbortController):
   await execBrew(cmd, cancel);
 }
 
+export async function brewCleanup(cancel?: AbortController): Promise<void> {
+  await execBrew(`cleanup --prune=all`, cancel);
+}
+
 export async function brewPinFormula(formula: Formula | OutdatedFormula): Promise<void> {
   await execBrew(`pin ${formula.name}`);
   formula.pinned = true;