+++ extensions/brew/src/clean-up.tsx
@@ -1,15 +1,14 @@
 import { showToast, Toast } from "@raycast/api";
-import { brewUpgradeAll } from "./brew";
-import { preferences } from "./preferences";
+import { brewCleanup } from "./brew";
 import { showActionToast, showFailureToast, wait } from "./utils";
 
 export default async (): Promise<void> => {
   try {
-    const abort = showActionToast({ title: "Upgrading formula & casks" + String.ellipsis, cancelable: true });
-    await brewUpgradeAll(preferences.greedyUpgrades, abort);
-    showToast(Toast.Style.Success, "Upgrade completed");
+    const abort = showActionToast({ title: "Cleaning up everything at once" + String.ellipsis, cancelable: true });
+    await brewCleanup(abort);
+    showToast(Toast.Style.Success, "Cleaning completed");
   } catch (err) {
-    await showFailureToast("Upgrade failed", err as Error);
+    await showFailureToast("Cleaning failed", err as Error);
     // Wait around until user has had chance to click the Toast action.
     // Note this only works for "no view" commands (actions still break when popping a view based command).
     // See: https://raycastapp.slack.com/archives/C01E6LWGXJ8/p1642676284027700